requirements.txt
absl-py==0.9.0
alembic==1.4.2
async-generator==1.10
attrs==19.3.0
autograd==1.3
autograd-gamma==0.5.0
azure-core==1.7.0
azure-storage-blob==12.3.2
backcall==0.1.0
beautifulsoup4==4.8.2
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
blinker==1.4
bokeh==1.4.0
branca==0.3.1
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
certipy==0.1.3
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.9.2
cycler==0.10.0
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1591799499719/work
cytoolz==0.10.1
dask==2.9.2
databricks-cli==0.11.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distributed @ file:///home/conda/feedstock_root/build_artifacts/distributed_1588976153416/work
docker==4.2.2
elasticsearch==7.1.0
entrypoints==0.3
fastcache==1.1.0
feather-format==0.4.1
Flask==1.1.2
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1589989738418/work
future==0.18.2
gitdb==4.0.5
GitPython==3.1.7
gmpy2==2.1.0b1
google-auth==1.17.0
google-auth-oauthlib==0.4.1
gorilla==0.3.0
graphviz==0.11
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
hiddenlayer==0.3
idna==2.9
imageio==2.8.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1591451751445/work
install==1.3.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1590020200501/work/dist/ipykernel-5.3.0-py3-none-any.whl
ipyleaflet @ file:///home/conda/feedstock_root/build_artifacts/ipyleaflet_1590404801797/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1590796899444/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1589812474002/work
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1591810480056/work
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-tabnine==1.0.2
jupyter-telemetry==0.0.5
jupyterhub==1.1.0
jupyterlab==1.2.5
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1590229434073/work
kaleido==0.0.3.post1
kiwisolver==1.2.0
lifelines==0.25.5
llvmlite==0.31.0
locket==0.2.0
lxml @ file:///home/conda/feedstock_root/build_artifacts/lxml_1589927514031/work
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mlflow==1.10.0
monai==0.1.0
mpmath==1.1.0
msgpack==1.0.0
msrest==0.6.18
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nibabel==3.1.0
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1588887226267/work
numba==0.48.0
numexpr==2.7.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1591485215893/work
oauthlib==3.0.1
olefile==0.46
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pamela==1.0.0
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.12.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prometheus-flask-exporter==0.15.4
prompt-toolkit==3.0.5
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycox @ file:///home/jovyan/pycox
pycparser==2.20
pycurl==7.43.0.5
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
pytorch-ignite==0.3.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rise==5.6.1
rpy2==2.9.4
rsa==4.2
ruamel-yaml==0.15.80
ruamel.yaml.clib==0.2.0
scikit-fuzzy==0.4.2
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
SimpleITK==1.2.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn-fuzzy==0.0.8
sklearn-pandas==2.0.0
smmap==3.0.4
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1591541743024/work
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1589778966114/work
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.11.1
sympy==1.5.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.10.0
torch==1.5.0
torchtuples==0.2.0
torchvision==0.2.2
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1591181521996/work
traitlets==4.3.3
traittypes==0.2.1
tzlocal @ file:///home/conda/feedstock_root/build_artifacts/tzlocal_1588939190034/work
urllib3==1.25.9
vincent==0.4.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1591600393557/work
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zict==2.0.0
zipp==3.1.0
